feat: enhance reMatchEquipment tests for CBR and DIS equipment reassignment scenarios
